Move CheckboxListOptionControl to its own file
