+++ plugins/woocommerce-blocks/packages/components/checkbox-list/types.ts
@@ -1,0 +1,4 @@
+export interface CheckboxListOptions {
+	label: React.ReactNode;
+	value: string;
+}
